Moved relevant code to settings service
